agent-driven: Extract parse_script_extensions to ucd_helper.rs
